refactor: extract AliceResourceUtilities into 3 focused classes (916→496 lines)
Extract model resource loading, texture/material handling, and resource enum
resolution into dedicated classes while preserving full backward compatibility
through 36 delegate methods in the slimmed facade.

New classes (all package-private):
- ResourceEnumResolver: enum↔camelCase conversion, resource name resolution
- ResourceTextureManager: filename construction, URL resolution
- ModelResourceLoader: binary codec, caching, skeleton operations

Key changes:
- AliceResourceUtilities: 916→496 lines (under 500 target)
- ResourceNames inner class + cache moved to ResourceEnumResolver
- urlToVisualMap/urlToTextureMap caches moved to ModelResourceLoader
- JsonModelIoTest reflection targets updated for ResourceEnumResolver
- All 220 existing tests pass without modification

Closes #580
